fixes #115, also limits A12/A13 downgrading to macOS 15 and later on macOS systems

+++ surrealra1n.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-CURRENT_VERSION="v2.0 beta 19"
+CURRENT_VERSION="v2.0 beta 20"
 
 if [ "$EUID" -eq 0 ]; then
   echo "ERROR: Do not run this script with sudo or as root."
@@ -503,18 +503,20 @@ elif [[ $dist == 3 ]]; then
     # install Kernel64Patcher for tether booting iOS 13+
     curl -L -o bin/Kernel64Patcher https://github.com/edwin170/downr1n/raw/refs/heads/main/binaries/Darwin/Kernel64Patcher
     # fetch pwnerblu fork of Kernel64Patcher and iBootpatch2 for tether booting iOS 14.x on A12 device.
-    git clone https://github.com/pwnerblu/Kernel64Patcher --recursive
-    cd Kernel64Patcher
-    make
-    cp Kernel64Patcher ../bin/Kernel64Patcher3
-    cd ..
-    rm -rf "Kernel64Patcher"
-    git clone https://github.com/pwnerblu/iBootpatch2 -b ipad6
-    cd iBootpatch2
-    make
-    cp iBootpatch2 ../bin/iBootpatch2
-    cd ..
-    rm -rf "iBootpatch2"
+    if [[ $macos_ver == 15.* || $macos_ver == 26.* || $macos_ver == 27.* ]]; then
+        git clone https://github.com/pwnerblu/Kernel64Patcher --recursive
+        cd Kernel64Patcher
+        make
+        cp Kernel64Patcher ../bin/Kernel64Patcher3
+        cd ..
+        rm -rf "Kernel64Patcher"
+        git clone https://github.com/pwnerblu/iBootpatch2 -b ipad6
+        cd iBootpatch2
+        make
+        cp iBootpatch2 ../bin/iBootpatch2
+        cd ..
+        rm -rf "iBootpatch2"
+    fi
     # done!
     curl -L -o bin/gaster https://github.com/LukeZGD/Legacy-iOS-Kit/raw/refs/heads/main/bin/macos/gaster
     curl -L -o bin/tsschecker https://github.com/LukeZGD/Legacy-iOS-Kit/raw/refs/heads/main/bin/macos/tsschecker
@@ -602,18 +604,20 @@ elif [[ $dist == 4 ]]; then
     # install Kernel64Patcher for tether booting iOS 13+
     curl -L -o bin/Kernel64Patcher https://github.com/edwin170/downr1n/raw/refs/heads/main/binaries/Darwin/Kernel64Patcher
     # fetch pwnerblu fork of Kernel64Patcher and iBootpatch2 for tether booting iOS 14.x on A12 device.
-    git clone https://github.com/pwnerblu/Kernel64Patcher --recursive
-    cd Kernel64Patcher
-    make
-    cp Kernel64Patcher ../bin/Kernel64Patcher3
-    cd ..
-    rm -rf "Kernel64Patcher"
-    git clone https://github.com/pwnerblu/iBootpatch2 -b ipad6
-    cd iBootpatch2
-    make
-    cp iBootpatch2 ../bin/iBootpatch2
-    cd ..
-    rm -rf "iBootpatch2"
+    if [[ $macos_ver == 15.* || $macos_ver == 26.* || $macos_ver == 27.* ]]; then
+        git clone https://github.com/pwnerblu/Kernel64Patcher --recursive
+        cd Kernel64Patcher
+        make
+        cp Kernel64Patcher ../bin/Kernel64Patcher3
+        cd ..
+        rm -rf "Kernel64Patcher"
+        git clone https://github.com/pwnerblu/iBootpatch2 -b ipad6
+        cd iBootpatch2
+        make
+        cp iBootpatch2 ../bin/iBootpatch2
+        cd ..
+        rm -rf "iBootpatch2"
+    fi
     # done!
     curl -L -o bin/gaster https://github.com/LukeZGD/Legacy-iOS-Kit/raw/refs/heads/main/bin/macos/gaster
     curl -L -o bin/tsschecker https://github.com/LukeZGD/Legacy-iOS-Kit/raw/refs/heads/main/bin/macos/tsschecker
@@ -2712,6 +2716,15 @@ exit 0
 }
 
 do_tethered_restore_a12_a13(){
+
+if [[ $macos_ver == 15.* || $macos_ver == 26.* || $macos_ver == 27.* ]] && [[ $dist == 3 || $dist == 4 ]]; then
+    echo ""
+elif [[ $dist != 3 && $dist != 4 ]]; then
+    echo ""
+else
+    echo "A12/A13 downgrades are only supported on macOS 15 and later."
+    exit 1
+fi
 
 if [[ -z "$IPSW_PATH" ]]; then
     echo "No IPSW selected. Aborting."
